requirements.txt
Django==5.2.7
djangorestframework==3.15.1
django-cors-headers==4.4.0
pandas==2.2.1
scikit-learn==1.4.1.post1
joblib==1.3.2
gunicorn==21.2.0
whitenoise==6.6.0
python-dotenv==1.0.0  # <-- Importante para .env